fix bugs

+++ app/Models/BaseModel.php
@@ -17,7 +17,7 @@ class BaseModel extends Model
     {
         parent::boot();
         if (Auth::check()) {
-            static::addGlobalScope(new OrganizationScope(Auth::user()->getOriginal("organization_id")));
+            static::addGlobalScope(new OrganizationScope((int)Auth::user()->getOriginal("organization_id")));
         } else {
             static::addGlobalScope(new OrganizationScope(1));
         }
